+++ media/webviewProjectScripts.js
@@ -82,9 +82,113 @@ function initProjects() {
         return true;
     }
 
-    document.addEventListener('click', function (e) {
+    var contextMenuProjectId = null;
+    var contextMenuGroupId = null;
+    function onContextMenu(e) {
+        closeContextMenus(); // Close previews
+
+        var projectDiv = e.target.closest('.project[data-id]');
+        var groupDiv = e.target.closest('.group-title')
+        if (!projectDiv && !groupDiv)
+            return;
+
+        e.preventDefault();
+
+        let contextMenuForProject = projectDiv != null;
+        var contextMenuElement;
+        if (contextMenuForProject) {
+            contextMenuProjectId = projectDiv.getAttribute("data-id");
+            if (contextMenuProjectId == null)
+                return;
+
+            contextMenuElement = document.getElementById("projectContextMenu");
+        } else {
+            let groupIdDiv = groupDiv.closest(".group[data-group-id]");
+            contextMenuGroupId = groupIdDiv ? groupIdDiv.getAttribute("data-group-id") : null;
+            if (contextMenuGroupId == null)
+                return;
+
+            contextMenuElement = document.getElementById("groupContextMenu");
+        }
+
+        contextMenuElement.style.left = e.clientX + "px";
+        contextMenuElement.style.top = e.clientY + "px";
+        contextMenuElement.classList.add("visible");
+    }
+
+    function onProjectContextMenuActionClicked(el) {
+        var action = el.getAttribute("data-action");
+
+        if (action == null || contextMenuProjectId == null)
+            return;
+
+        switch (action) {
+            case 'open':
+                onProjectClicked(contextMenuProjectId, false);
+                break;
+            case 'open-new-window':
+                onProjectClicked(contextMenuProjectId, true);
+                break;
+            default:
+                window.vscode.postMessage({
+                    type: action + '-project',
+                    projectId: contextMenuProjectId,
+                });
+                break;
+        }
+
+        closeContextMenus();
+    }
+
+    function onGroupContextMenuActionClicked(el) {
+        var action = el.getAttribute("data-action");
+
+        if (action == null || contextMenuGroupId == null)
+            return;
+
+        switch (action) {
+            case 'add':
+                window.vscode.postMessage({
+                    type: 'add-project',
+                    groupId: contextMenuGroupId,
+                });
+                break;
+            default:
+                window.vscode.postMessage({
+                    type: action + '-group',
+                    groupId: contextMenuGroupId,
+                });
+                break;
+        }
+
+        closeContextMenus();
+    }
+
+    function closeContextMenus() {
+        contextMenuProjectId = null;
+        contextMenuGroupId = null;
+        document.querySelectorAll(".custom-context-menu").forEach(element =>
+            element.classList.remove("visible")
+        );
+    }
+
+    document.addEventListener('click', (e) => {
         if (!e.target)
             return;
+
+        var contextMenuElement = e.target.closest("#projectContextMenu [data-action]");
+        if (contextMenuElement) {
+            onProjectContextMenuActionClicked(contextMenuElement);
+            return;
+        }
+
+        contextMenuElement = e.target.closest("#groupContextMenu [data-action]");
+        if (contextMenuElement) {
+            onGroupContextMenuActionClicked(contextMenuElement);
+            return;
+        }
+
+        closeContextMenus();
 
         if (e.target.closest('[data-action="add-group"]')) {
             window.vscode.postMessage({
@@ -111,4 +215,17 @@ function initProjects() {
         .forEach(element =>
             element.addEventListener("click", onAddProjectClicked)
         );
+
+    document.addEventListener('contextmenu', (e) => {
+        if (!e.target)
+            return;
+
+        onContextMenu(e);
+    });
+
+    document.addEventListener("keydown", e => {
+        if (e.key === "Escape") {
+            closeContextMenus();
+        }
+    });
 }